Update to use Elasticsearch fitsheaders alias.
The LCO fitsheaders ES index now has an alias, lco-fitsheaders, which all client software should point to. This improves the sustainability of the index, namely when re-indexing to change the schema as necessary. Re-indexing requires the creation of a new index and a copy operation. When addressing an alias, this migration process will be transparent all client software using said alias. See https://www.lewuathe.com/how-to-reindex-elasticsearch.html for more details.